+++ src/sqlfluff/core/parser/segments/base.py
@@ -70,10 +70,8 @@ class BaseSegment:
     # We define this as Null here but it is assumed that any subclass must override.
     match_grammar: Matchable = None  # type: ignore
     comment_seperate = False
-    is_whitespace = False
     optional = False  # NB: See the sequence grammar for details
-    is_segment = True
-    _name = None
+    _name: Optional[str] = None
     is_meta = False
     # Are we able to have non-code at the start or end?
     can_start_end_non_code = False
@@ -204,6 +202,11 @@ class BaseSegment:
     def is_comment(self):
         """Return True if this is entirely made of comments."""
         return all(seg.is_comment for seg in self.segments)
+
+    @cached_property
+    def is_whitespace(self):
+        """Return True if this segment is entirely whitespace."""
+        return all(seg.is_whitespace for seg in self.segments)
 
     @cached_property
     def raw(self):