+++ src/sqlfluff/dialects/dialect_exasol.py
@@ -8,6 +8,7 @@ from sqlfluff.core.parser import (
     AnyNumberOf,
     BaseSegment,
     Bracketed,
+    OptionallyBracketed,
     Dedent,
     Delimited,
     GreedyUntil,
@@ -20,6 +21,10 @@ from sqlfluff.core.parser import (
     ReSegment,
     Sequence,
     StartsWith,
+    RegexMatcher,
+    StringMatcher,
+    CodeSegment,
+    CommentSegment,
 )
 from sqlfluff.core.dialects import load_raw_dialect
 from sqlfluff.dialects.exasol_keywords import (
@@ -39,27 +44,27 @@ exasol_dialect.sets("reserved_keywords").update(RESERVED_KEYWORDS)
 exasol_dialect.sets("bare_functions").clear()
 exasol_dialect.sets("bare_functions").update(BARE_FUNCTIONS)
 
-exasol_dialect.set_lexer_struct(
+exasol_dialect.insert_lexer_matchers(
     [
-        ("range_operator", "regex", r"\.{2}", dict(is_code=True)),
-        ("hash", "singleton", "#", dict(is_code=True)),
-    ]
-    + exasol_dialect.get_lexer_struct()
+        RegexMatcher("range_operator", r"\.{2}", CodeSegment),
+        StringMatcher("hash", "#", CodeSegment),
+    ],
+    before="not_equal",
 )
 
-exasol_dialect.patch_lexer_struct(
+exasol_dialect.patch_lexer_matchers(
     [
         # In EXASOL, a double single/double quote resolves as a single/double quote in the string.
         # It's also used for escaping single quotes inside of STATEMENT strings like in the IMPORT function
         # https://docs.exasol.com/sql_references/basiclanguageelements.htm#Delimited_Identifiers
         # https://docs.exasol.com/sql_references/literals.htm
-        ("single_quote", "regex", r"'([^']|'')*'", dict(is_code=True)),
-        ("double_quote", "regex", r'"([^"]|"")*"', dict(is_code=True)),
-        (
+        RegexMatcher("single_quote", r"'([^']|'')*'", CodeSegment),
+        RegexMatcher("double_quote", r'"([^"]|"")*"', CodeSegment),
+        RegexMatcher(
             "inline_comment",
-            "regex",
             r"--[^\n]*",
-            dict(is_comment=True, type="comment", trim_start=("--")),
+            CommentSegment,
+            segment_kwargs={"trim_start": ("--")},
         ),
     ]
 )
@@ -1322,7 +1327,7 @@ class UpdateStatementSegment(BaseSegment):
     match_grammar = StartsWith("UPDATE")
     parse_grammar = Sequence(
         "UPDATE",
-        Ref("AliasedTableReferenceSegment"),
+        OneOf(Ref("TableReferenceSegment"), Ref("AliasedTableReferenceGrammar")),
         Ref("SetClauseListSegment"),
         Ref("UpdateFromClauseSegment", optional=True),
         Ref("WhereClauseSegment", optional=True),
@@ -1375,7 +1380,7 @@ class UpdateFromClauseSegment(BaseSegment):
     match_grammar = Sequence(
         "FROM",
         Delimited(
-            Ref("AliasedTableReferenceSegment"),
+            OneOf(Ref("TableReferenceSegment"), Ref("AliasedTableReferenceGrammar")),
             terminator="WHERE",
         ),
     )
@@ -1404,7 +1409,7 @@ class MergeStatementSegment(BaseSegment):
     parse_grammar = Sequence(
         "MERGE",
         "INTO",
-        Ref("AliasedTableReferenceSegment"),
+        OneOf(Ref("TableReferenceSegment"), Ref("AliasedTableReferenceGrammar")),
         "USING",
         OneOf(
             Ref("TableReferenceSegment"),  # tables/views
@@ -1527,7 +1532,7 @@ class DeleteStatementSegment(BaseSegment):
         "DELETE",
         Ref("StarSegment", optional=True),
         "FROM",
-        Ref("AliasedTableReferenceSegment"),
+        OneOf(Ref("TableReferenceSegment"), Ref("AliasedTableReferenceGrammar")),
         Ref("WhereClauseSegment", optional=True),
         Ref("PreferringClauseSegment", optional=True),
     )
@@ -2477,10 +2482,7 @@ class PreferringClauseSegment(BaseSegment):
     )
     parse_grammar = Sequence(
         "PREFERRING",
-        OneOf(
-            Ref("PreferringPreferenceTermSegment"),
-            Bracketed(Ref("PreferringPreferenceTermSegment")),
-        ),
+        OptionallyBracketed(Ref("PreferringPreferenceTermSegment")),
         Ref("PartitionClauseSegment", optional=True),
     )