+++ src/sqlfluff/dialects/dialect_teradata.py
@@ -19,16 +19,18 @@ from sqlfluff.core.parser import (
     AnyNumberOf,
     Ref,
     Anything,
+    RegexMatcher,
+    CodeSegment,
 )
 from sqlfluff.core.dialects import load_raw_dialect
 
 ansi_dialect = load_raw_dialect("ansi")
 teradata_dialect = ansi_dialect.copy_as("teradata")
 
-teradata_dialect.patch_lexer_struct(
+teradata_dialect.patch_lexer_matchers(
     [
-        # name, type, pattern, kwargs, so it also matches 1.
-        ("numeric_literal", "regex", r"([0-9]+(\.[0-9]*)?)", dict(is_code=True)),
+        # so it also matches 1.
+        RegexMatcher("numeric_literal", r"([0-9]+(\.[0-9]*)?)", CodeSegment),
     ]
 )
 
@@ -485,9 +487,7 @@ class CreateTableStatementSegment(BaseSegment):
                         delimiter=Ref("CommaSegment"),
                     )
                 ),
-                Sequence(  # [COMMENT 'string'] (MySQL)
-                    "COMMENT", Ref("QuotedLiteralSegment"), optional=True
-                ),
+                Ref("CommentClauseSegment", optional=True),
             ),
             # Create AS syntax:
             Sequence("AS", Ref("SelectableGrammar")),