+++ test/fixtures/linter/operator_errors_ignore.sql
@@ -5,7 +5,7 @@ are still present. No errors should be found on line 8 at all. */
 
 SELECT
     a.a + a.b AS good,
-    a.a  - a.b AS bad_1,  -- noqa
-    a.a *  a.b AS bad_2,  -- noqa: L007, L006
+    a.a-a.b AS bad_1,  -- noqa
+    a.a*a.b AS bad_2,  -- noqa: L007, L006
     a.a*a.b AS bad_3  -- noqa: L007
 FROM tbl AS a